fixing python tests
